docs: clarify Codex flags usage
